o  deprecated getObjectByID() methods that use 'int' as an Id. Use version with 'long' o  related updates to unit tests.

+++ plexus-components/trunk/plexus-jdo2/src/test/java/org/codehaus/plexus/jdo/Child.java
@@ -37,7 +37,7 @@ public class Child
     /**
      * Field id
      */
-    private int id = 0;
+    private long id = 0;
 
     /**
      * Field name
@@ -83,7 +83,7 @@ public class Child
     /**
      * Get id
      */
-    public int getId()
+    public long getId()
     {
         return this.id;
     }
@@ -121,7 +121,7 @@ public class Child
      * 
      * @param id
      */
-    public void setId( int id )
+    public void setId( long id )
     {
         this.id = id;
     }
